수정 : 리뷰목록 개선 (지정개수 출력 및 정보변경)

+++ reservationweb/src/main/java/or/connect/reservationweb/dao/ReservationCommentDao.java
@@ -53,7 +53,7 @@ public class ReservationCommentDao {
 						commentDto.setCommentId(rs.getInt("commentId"));
 						commentDto.setProductId(rs.getInt("productId"));
 						commentDto.setReservationInfoId(rs.getInt("reservationInfoId"));
-						commentDto.setScore(rs.getInt("score"));
+						commentDto.setScore(rs.getDouble("score"));
 						commentDto.setComment(rs.getString("comment"));
 						commentDto.setReservationName(rs.getString("reservation_name"));
 						commentDto.setReservationTelephone(rs.getString("reservation_tel"));
